moved tests for edit into its own module
